Refactor test file `rules/__tests__/index`
- Rewrite to ESM (`.mjs`)
- Use async/await as possible
- Include missing `stylelint-config-recommended`
